+++ tokens/usecase/pricing/worker/liquidity_pricer.go
@@ -12,20 +12,92 @@ import (
 type liquidityPricer struct {
 	defaultQuoteDenom string
 
-	quoteDenomScalingFactor osmomath.BigDec
+	scalingFactorGetterCb domain.ScalingFactorGetterCb
 }
+
+const liquidityCapErrorSeparator = "; "
 
 var _ domain.LiquidityPricer = &liquidityPricer{}
 
-func NewLiquidityPricer(defaultQuoteDenom string, quoteDenomScalingFactor osmomath.Dec) domain.LiquidityPricer {
+func NewLiquidityPricer(defaultQuoteDenom string, chainScalingFactorGetterCb domain.ScalingFactorGetterCb) domain.LiquidityPricer {
 	return &liquidityPricer{
-		defaultQuoteDenom:       defaultQuoteDenom,
-		quoteDenomScalingFactor: osmomath.BigDecFromDec(quoteDenomScalingFactor),
+		defaultQuoteDenom: defaultQuoteDenom,
+
+		scalingFactorGetterCb: chainScalingFactorGetterCb,
 	}
 }
 
-// ComputeCoinCap implements LiquidityPricer.
-func (l *liquidityPricer) ComputeCoinCap(coin sdk.Coin, baseDenomPriceData domain.DenomPriceInfo) (math.LegacyDec, error) {
+// PriceCoin implements domain.PoolLiquidityPricerWorker.
+func (p *liquidityPricer) PriceCoin(coin sdk.Coin, price osmomath.BigDec) osmomath.Int {
+	if price.IsZero() {
+		// If the price is zero, set the capitalization to zero.
+		return osmomath.ZeroInt()
+	}
+
+	// Get the scaling factor for the base denom.
+	baseScalingFactor, err := p.scalingFactorGetterCb(coin.Denom)
+	if err != nil {
+		// If there is an error, keep the total liquidity but set the capitalization to zero.
+		return osmomath.ZeroInt()
+	}
+
+	priceInfo := domain.DenomPriceInfo{
+		Price:         price,
+		ScalingFactor: baseScalingFactor,
+	}
+
+	liquidityCapitalization, err := ComputeCoinCap(coin, priceInfo)
+	if err != nil {
+		// If there is an error, keep the total liquidity but set the capitalization to zero.
+		return osmomath.ZeroInt()
+	}
+
+	return liquidityCapitalization.TruncateInt()
+}
+
+// PriceBalances implements domain.PoolLiquidityPricerWorker.
+func (p *liquidityPricer) PriceBalances(balances sdk.Coins, prices domain.PricesResult) (osmomath.Int, string) {
+	totalCapitalization := osmomath.ZeroInt()
+
+	// Note: errors may occur in any denom.
+	// As a result, we accumulate them in this error string
+	// to ease debugging if issues occur.
+	liquidityCapErrorStr := ""
+
+	for _, balance := range balances {
+		denom := balance.Denom
+
+		price := prices.GetPriceForDenom(denom, p.defaultQuoteDenom)
+
+		currentCapitalization := p.PriceCoin(balance, price)
+
+		if currentCapitalization.IsZero() {
+			if len(liquidityCapErrorStr) != 0 {
+				liquidityCapErrorStr += liquidityCapErrorSeparator
+			}
+
+			liquidityCapErrorStr += formatLiquidityCapErrorStr(denom)
+		}
+
+		totalCapitalization = totalCapitalization.Add(currentCapitalization)
+	}
+
+	return totalCapitalization, liquidityCapErrorStr
+}
+
+// formatLiquidityCapErrorStr formats the liquidity cap error
+func formatLiquidityCapErrorStr(denom string) string {
+	return fmt.Sprintf("zero cap for denom (%s)", denom)
+}
+
+// ComputeCoinCap computes the equivalent of the given coin in the desired quote denom that is set on ingester.
+//
+// Returns error if:
+// * Price is zero
+// * Scaling factor is zero
+// * Truncation occurs in intermediary operations. Truncation is defined as the original amount
+// being non-zero and the computed amount being zero.
+func ComputeCoinCap(coin sdk.Coin, baseDenomPriceData domain.DenomPriceInfo) (math.LegacyDec, error) {
 	if baseDenomPriceData.Price.IsZero() {
 		return osmomath.Dec{}, fmt.Errorf("price for %s is zero", coin.Denom)
 	}
@@ -33,11 +105,10 @@ func (l *liquidityPricer) ComputeCoinCap(coin sdk.Coin, baseDenomPriceData domai
 		return osmomath.Dec{}, fmt.Errorf("scaling factor for %s is zero", coin.Denom)
 	}
 
-	currentCoinCap := osmomath.BigDecFromSDKInt(coin.Amount).MulMut(baseDenomPriceData.Price)
+	currentCoinCap := osmomath.BigDecFromSDKInt(coin.Amount).MulMut(baseDenomPriceData.Price).QuoMut(osmomath.BigDecFromDec(baseDenomPriceData.ScalingFactor))
 	isOriginalAmountZero := coin.Amount.IsZero()
 
 	// Truncation in intermediary operation - return error.
-	currentCoinCap = l.quoteDenomScalingFactor.Mul(currentCoinCap).QuoMut(osmomath.BigDecFromDec(baseDenomPriceData.ScalingFactor))
 	if currentCoinCap.IsZero() && !isOriginalAmountZero {
 		return osmomath.Dec{}, fmt.Errorf("truncation occurred when multiplying (%s) of denom (%s) by the scaling factor (%s)", currentCoinCap, coin.Denom, baseDenomPriceData.ScalingFactor)
 	}